Humidity

+++ Api/Models/HumidityMeasurement.cs
@@ -6,7 +6,7 @@ namespace Api.Models
     public class HumidityMeasurement
     {
         public string GreenHouseId { get; set; }
-        public int Humidity { get; set; }
+        public double Humidity { get; set; }
         public long Time { get; set; }
         
     }
